requirements.txt
requests
python-dotenv
nltk
pyperclip
streamlit
streamlit-auth0